Remove rate_estimate check before inferring TCP RTT
The eBPF code for inferring the TCP RTT contained a check that the
estimated throughput in at least one direction of the flow must be
non-zero. As the rate estimation is only updated at most once per
second, and is initialized to zero, this will prevent any RTTs from
flows that started less than a second ago (or that after a quiet
period has not sent data for at least 1 second). Consequently, no RTTs
from short flows will be captured, nor will any RTTs from the earlier
part of the flow (e.g. when some initial bi-directional communication
may occur and before the flow starts to ramp up its throughput) be
included. As the check is just for non-zero throughput though, RTTs
from very low traffic flows may still be included.

Remove this rate check, to prevent RTTs from all short flows being
missed, and from RTTs during the earlier phase of the connection being
ignored.

If the intention of the check is to explicitly avoid RTTs from short
flows (or the early parts of long flows), it's better to add a check
against the start time (data->start_time) of the flow to more clearly
express the intent, rather then relying on non-updated rate
estimations. If the intention is to avoid low-throughput flows, a
different limit that non-zero should be used, and ideally the estimate
should be updated quickly after the flow start. If the intention is to
generally avoid too many RTTs being pushed to user space, it may be
better to put a rate-limit on how often the RTTs can be reported,
rather than preventing all RTTs from short flows (while allowing
unlimited number of RTTs for long elephant flows) to reduce bias in
the type of flows RTTs are collected from.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>