Do not track TSval for pure ACKs
Ensure that we only record TSvals (that TSecrs will later be matched
again) for TCP packets carrying payload (or SYN packets). While pure
ACKs (i.e. TCP packets with no payload, only ACKing received data) can
have new TSvals, we should NOT attempt to match a TSecr against those,
as that might result in a highly inflated RTT value.

Consider the scenario with a "unidirectional" TCP flow, i.e. data is
only sent in one direction, with only pure ACKs being sent in the
opposite direction. Matching the TSval on the data packets from the
TCP sender with the TSecr in the ACKs from the TCP receiver will
produce valid RTT samples, as the TCP receiver must ACK the data (and
then echo the TSval as a TSecr) in a timely manner.

However, matching the TSval from one of the ACKs from the TCP receiver
with a TSecr in a later data packet from the TCP sender, is not
guaranteed to produce a valid RTT sample. While the TCP sender will
echo the TSval from the ACK as a TSecr in the next packet it sends,
the TCP sender may wait an arbitrary time before actually sending the
next packet. If the TCP sender waits for a long time, e.g. due to the
application only periodically transmitting data, this can create a
large inflation in the RTT value when the TSecr matching the TSval in
the ACK finally arrives.

In general, any RTT sample obtained by matching a TSecr to a TSval
from a pure ACK will consist of the network RTT + d, where d is the
delay between the sender receiving the ACK and sending a new
packet. While the RTT sample may be useful in those cases where d is
sufficiently low (e.g. because TCP sender is constantly sending packets
at a high rate), we do not generally know when d is low or not.

It would be possible to create an upper bound for d by keeping track
of the inter-packet delay for the TCP sender, and then only accept the
RTT sample if the inter-packet delay is low enough to ensure that d
does not contribute much to the overall RTT. However this adds a fair
bit of additional complexity, so instead just avoid creating TSval
records that we can match against for pure ACKs. Note that we still
update the last-seen TSval even for pure ACKs, so that we are still
able to detect (and only track) new TSval values.

Implementation wise, the get_tcp_segment_size() helper might
conceptually fit better as part of dissector.h (as part of its snoop()
function). However, currently the TCP payload size is only needed here
to avoid creating inflated RTT samples, so place it closely to where
it is used. This way, we also only pay the overhead for calculating
the TCP segment size when need it, although the overhead of actually
doing that calculation will be slightly higher as we need to redo some
boundary checks to please the verifier.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>