Make updated RTT tracking compatible with v5.15 kernel
The changes introduced to the eBPF RTT tracking logic in the previous
commits, unfortunately made the logic too complex for the eBPF
verifier in the v5.15 kernel (hits the 1 million instruction limit).

Through some trail and error I found that checking that the TCP
segment size is non-zero (not a pure ACK) had a large impact on the
overall complexity (simply dropping the check reduces the verification
complexity from over 1m to just over 530k instructions).

To reduce the complexity of computing the TCP segment size, move the
logic directly into the dissector, which avoids having to redo some
boundary checks to please the verifier. This is sufficient to reduce
the verification complexity down to 956k (i.e. just barley within the
limits).

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>