Attempt move dissector struct off stack
The dissector_t struct is a large (112 bytes) struct that's kept on
stack. Reduce stack usage by moving this struct into a per-CPU
variable (which should be safe as XDP programs are not reentrant).

The (5.15) verifier currently refuses the load this program though as
it no longer seems to treat the dissector->start and dissector->end
pointers as packet pointers. Trying to convince the verifier that this
is still the case might be very challenging and require changes across
the entire XDP program (dissector, flows, heimdall etc). So will
likely give up on this approach.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>