Add some debug counters and printouts to original RTT code
Can unfortunately not branch this of the original develop branch, as
there is not enough stack space left over there for the if-statements
to be processed. Therefore I'm basing this on top of the
stack-reducing commits.

Obviously this is not valid for a performance comparison, only used to
asses number of RTTs and their accuracy.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>